Don't rely on margins being shared between pages
We were saving screen content to a background page while margins were
set, and later restoring that content after the margins had been reset.
This relied on the margins being shared between pages, but that isn't
the case on a real terminal, so the saved content wasn't where we were
expecting it to be. This has now been fixed by having the margins reset
for both the save and restore steps.

Fixes #1