Fix for Save As dialog when opening a file via command line
When the app was started with a font file passed on the command line,
it would then be impossible to save any changes to a different name,
because the Save As dialog wouldn't open. This was just because we
didn't have a full file path, so we now append the current directory
to the file name that we're given.